requirements.txt
websocket-client==1.9.0
pip_system_certs==5.3
rel==0.4.9.22
SQLAlchemy==2.0.44